refactored MatrixAggregator to disconnect from MatrixCalculator, add task decorator for matrix related Configurables
