+++ src/vistiq/segment/__init__.py
@@ -44,8 +44,6 @@ from vistiq.segment.select import (
     FilterOpsConfig,
     MatrixFilter,
     MatrixFilterConfig,
-    MatrixAggregator,
-    MatrixAggregatorConfig,
     MaxFilter,
     MaxFilterConfig,
     MinFilter,
@@ -105,8 +103,6 @@ __all__ = [
     "LocalThresholdConfig",
     "MatrixFilter",
     "MatrixFilterConfig",
-    "MatrixAggregator",
-    "MatrixAggregatorConfig",
     "MaxFilter",
     "MaxFilterConfig",
     "MinFilter",
